Pushing changes to save state over routing, part 1
