+++ src/App.js
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useState } from "react";
 import NavBar from "./NavBar";
 import Home from "./Home";
 import Teams from "./Teams";
@@ -7,6 +7,30 @@ import * as apis from "./scripts/apis";
 import TeamInfo from "./TeamInfo";
 
 function App() {
+  const [trackedTeams, setTrackedTeams] = useState([]);
+
+  const teamTracker = {
+    getTeams: () => {
+      return trackedTeams;
+    },
+    containsTeam: teamString => {
+      return trackedTeams.includes(teamString);
+    },
+    addTeam: teamString => {
+      if (!trackedTeams.includes(teamString)) {
+        trackedTeams.push(teamString);
+        setTrackedTeams(trackedTeams);
+      }
+    },
+    removeTeam: teamString => {
+      const index = trackedTeams.indexOf(teamString);
+      if (index >= 0) {
+        trackedTeams.splice(index, 1);
+        setTrackedTeams(trackedTeams);
+      }
+    }
+  };
+
   return (
     <Router>
       <NavBar />
@@ -19,14 +43,22 @@ function App() {
         {/*<Route path="/Match-history/" component={MatchHistory} exact></Route>*/}
         <Route path="/Teams/" component={Teams} exact></Route>
         {/*<Route path="/" render={props => <Home {...props} />}></Route>*/}
-        <Route path="/" component={Home} exact></Route>
+        <Route
+          path="/"
+          render={props => <Home teamTracker={teamTracker} />}
+          exact
+        ></Route>
 
         {/* These routes are in the teams class*/}
         <Route path="/Teams/football" component={Teams} exact></Route>
         <Route path="/Teams/basketball" component={Teams} exact></Route>
         <Route path="/Teams/hockey" component={Teams} exact></Route>
         <Route path="/Teams/football/:team" component={TeamInfo} exact></Route>
-        <Route path="/Teams/basketball/:team" component={TeamInfo} exact></Route>
+        <Route
+          path="/Teams/basketball/:team"
+          component={TeamInfo}
+          exact
+        ></Route>
         <Route path="/Teams/hockey/:team" component={TeamInfo} exact></Route>
       </Switch>
     </Router>
